improved dashboard
- made suggested adjustments
- added icons to statcards
- added stylised result display for matches
- made existing icons into svg files for better resolution
- added numbers to seasonresult
