+++ frontend/src/app/page.jsx
@@ -3,9 +3,7 @@ import BarChart from "../components/SeasonGraph.jsx";
 import Matches from "../components/MatchCollection.jsx";
 
 
-export default function Home() {
-
-      let match_res=[
+let match_res=[
         {
             "id": "5ab46e31-391c-46a7-8e45-db9ada07626d",
             "result": "win",
@@ -35,24 +33,98 @@ export default function Home() {
         }
     ]
 
+export default function Dashboard() {
+  
   return (
-    <>
-      <StatCard
-        title={"Sejrsrate"}
-        body={"64.3%"}
-        footer={"18V - 7T - 3U"}
-        icon={"/icon1.jpg"}
-        iconAlt={"Win rate icon"}
-      />
-      <BarChart 
-        labels={['Sejre', 'Nederlag', 'Uafgjort']}
-        data={[18, 7, 3]}
-        bdcolor={['rgb(75, 156, 120)', 'rgb(209, 86, 80)', 'rgb(155, 160, 172)']}
-        bgcolor={['rgb(75, 156, 120, 1)', 'rgb(209, 86, 80, 1)', 'rgb(155, 160, 172, 1)']}
-        title={'Sæsonresultat'}
-      />
-      <Matches data={match_res}>
-      </Matches>
-    </>
+    <html 
+    lang="en"
+    className="__variable_188709 __variable_9a8899 h-full antialiased"
+    >
+        <head>
+            <title>Coach Dashboard</title>
+        </head>
+        <body className="min-h-full flex flex-col vc-init">
+            <header className="ballData-header">
+                <div>
+                    <h3>ballData</h3>
+                    <a style={{color:"var(--muted)"}}>Sæson {CurrentSeasonString()}</a>
+                </div>
+                <div style={{marginLeft:"auto", alignSelf:"center"}}>
+                    <a style={{color:"var(--hint)"}}>Sidst opdateret: {UpdateTime(new Date())}</a> {/*new date() so far, but needs to be changed when data actually gets updated*/}
+                </div>
+            </header>
+            <div className="dashboard-container">
+                <div className="statCards">
+                    <StatCard
+                            title={"Sejrsrate"}
+                            body={"64.3%"}
+                            footer={"18V - 7T - 3U"}
+                            iconColor={"#D0F0E6"}
+                            icon={"/trophy.svg"}
+                            iconAlt={"Win rate icon"}
+                        />
+                    <StatCard
+                            title={"Skudpræcision"}
+                            body={"75.7%"}
+                            footer={"Gns. 40.5 skud på mål pr. kamp"}
+                            iconColor={"#EAF1FB"}
+                            icon={"/target.svg"}
+                            iconAlt={"Win rate icon"}
+                        />
+                    <StatCard
+                            title={"Målmandsrednings\u{000AD}procent"} //\u{000AD} is a soft hyphen that suggests where to hyphenate the word if it is too long for the textbox
+                            body={"28.1%"}
+                            footer={"Gns. 19.0 redninger pr. kamp"}
+                            iconColor={"#FEF3DA"}
+                            icon={"/shield.svg"}
+                            iconAlt={"Win rate icon"}
+                        />
+                    <StatCard
+                            title={"Aktive spillere"}
+                            body={"18"}
+                            footer={"2 nye denne sæson"}
+                            iconColor={"#F3E8FF"}
+                            icon={"/active-players.svg"}
+                            iconAlt={"Win rate icon"}
+                        />
+                </div>
+                <div className="matchCollection-statistics">
+                    <Matches data={match_res}>
+                        </Matches>
+                        <BarChart 
+                                labels={['Sejre', 'Nederlag', 'Uafgjort']}
+                                data={[18, 7, 3]}
+                                bdcolor={['rgb(75, 156, 120)', 'rgb(209, 86, 80)', 'rgb(155, 160, 172)']}
+                                bgcolor={['rgb(75, 156, 120, 1)', 'rgb(209, 86, 80, 1)', 'rgb(155, 160, 172, 1)']}
+                                title={'Sæsonresultat'}
+                              />
+                </div>
+            </div>
+        </body>
+    </html>
   );
+}
+
+function CurrentSeasonString()
+{
+    let currentDate = new Date();
+    
+    //new season starts in september as far as I could find
+    return 8 <= currentDate.getMonth() ? 
+        `${currentDate.getFullYear()}/${currentDate.getFullYear() + 1}` : 
+        `${currentDate.getFullYear() - 1}/${currentDate.getFullYear()}`;
+}
+
+function UpdateTime(updateDate)
+{
+    const weekDays = ["Mandag", "Tirsdag", "Onsdag", "Torsdag", "Fredag", "Lørdag", "Søndag"];
+    const currentDate = new Date();
+    let updateDateString = updateDate.getDay() === currentDate.getDay() ? "I dag" : weekDays[updateDate.getDay()];
+
+    let updateTimeString = `${updateDate.getHours() + updateDate.getTimezoneOffset() / 60}:`; 
+    updateTimeString += updateDate.getMinutes() <= 9 ?
+        `0${updateDate.getMinutes()}` : 
+        updateDate.getMinutes();
+
+    return `${updateDateString}, ${updateTimeString}`;
 }
