+++ frontend/src/components/MatchRow.jsx
@@ -8,7 +8,11 @@ export default function MatchRow({data}) {
               <article className={styles.matchCard} key={info.id}>
                 <div className={styles.container}>
                     <div className={styles.icondiv}>
-                        {info.result && <Image src={info.result === 'win' ? "/win.jpg" : (info.result === 'loss' ? "/icon1.jpg" : "/win.jpg")} width={60} height={35} alt={info.iconAlt} />}
+                        {info.result && <div className={styles.resultDiv} style={{
+                            backgroundColor: info.result === "win" ? "#D0F0E6" : (info.result === "loss" ? "#FCEBEB" : "#EAF1FB"), 
+                            color: info.result === "win" ? "#1D9E75" : (info.result === "loss" ? "#E24B4A" : "#1D3A6E")}}>
+                                <>{info.result}</>
+                            </div>}
                     </div>
                     <div className={styles.textdiv}>
                         <p className={styles.locText}>{info.loc}</p>
