+++ frontend/src/components/styles.module.css
@@ -1,36 +1,41 @@
 .card {
-  width: 220px;
+  width: 1fr;
   background: white;
   border-radius: 14px;
   border: 1.5px solid var(--rule);
-  padding: 18px 20px 22px;
+  padding: 30px 18px 30px 22px;
   display: flex;
   flex-direction: column;
   position: relative;
 }
 
-.card img {
-  width: 54px;
-  height: 54px;
+.iconContainer{
+  width: 72px;
+  height: 72px;
   border-radius: 10px;
   background: var(--surface);
-  position: absolute;
-  top: 18px;
-  right: 18px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.iconContainer img {
+  width: 55%;
+  height: 55%;
 }
 
 .cardTitle {
   display: block;
   font-family: var(--sans);
-  font-size: 11px;
+  font-size: 14px;
   font-weight: 500;
   color: var(--hint);
   letter-spacing: 0.06em;
   text-transform: uppercase;
   margin: 0 0 14px;
-  width: 70%;
+  width: auto;
   overflow-wrap: break-word;
-  hyphens: manual;
+  hyphens: auto;
 }
 
 .cardBody {
@@ -42,7 +47,6 @@
   letter-spacing: -1px;
   margin: 0 0 18px;
   padding-bottom: 16px;
-  border-bottom: 1px solid var(--rule);
 }
 
 .cardFooter {
@@ -56,13 +60,12 @@
 .graphCard {
   display: flex;
   flex-direction: column;
-  width: 300px;
-  height: 250px;
+  width: auto;
+  height: 425px;
   background: white;
   border-radius: 14px;
   border: 1.5px solid var(--rule);
-  padding: 18px 20px 22px;
-  position: relative;
+  padding: 18px 0 18px;
 }
 
 .graphTitle {
@@ -73,8 +76,30 @@
   color: var(--ink);
   line-height: 1;
   letter-spacing: -1px;
-  padding-bottom: 16px;
+  padding-bottom: 32px;
   border-bottom: 1px solid var(--rule);
+  padding-left: 22px;
+  margin-bottom: 3%;
+}
+
+.graphNumbersContainer{
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  text-align: center;
+  margin-top: 25px;
+  font-family: "Space Grotesk";
+  font-weight: 400;
+  font-size: 24px;
+}
+
+.graphDescriptorsContainer{
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  text-align: center;
+  font-family: "Inter";
+  font-weight: 400;
+  font-size: 12px;
+  color: #5A6170;
 }
 
 .icondiv {
@@ -82,21 +107,33 @@
 }
 
 .textdiv {
-  float: right;
-  padding-left:10px; 
+  margin-left:10px;
 }
 
 .locText {
-  display: block;
   font-family: var(--sans);
   font-size: 14px;
   color: var(--hint);
+}
+
+.resultDiv{
+  width: 47px;
+  height: 35px;
+  padding: 0 10px 0 10px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border-radius: 8px;
+  font-family: "Space Grotesk";
+  text-transform: uppercase;
 }
 
 .container {
   display: flex;
   align-items: center;
 }
+
+
 
 .teamText {
   text-align: left;
@@ -126,8 +163,8 @@
 .collCard{
   display: flex;
   flex-direction: column;
-  width: 600px;
-  height: 600px;
+  width: 100%;
+  height: auto;
   background: white;
   border: 1.5px solid var(--rule);
   position: relative;
